+++ Sources/ArknightsClient/Features/Launcher/Settings/SettingsStrings.swift
@@ -137,10 +137,6 @@ enum SettingsStrings {
 		.settingsInstallationChinaClients
 	static let chinaClientsDetail = LocalizedStringResource.Settings
 		.settingsInstallationChinaClientsDetail
-	static let runtimePerformance = LocalizedStringResource.Settings
-		.settingsInstallationRuntimePerformance
-	static let runtimePerformanceDetail = LocalizedStringResource.Settings
-		.settingsInstallationRuntimePerformanceDetail
 	static let frameLatency = LocalizedStringResource.Settings.settingsInstallationFrameLatency
 	static let frameLatencyDetail = LocalizedStringResource.Settings
 		.settingsInstallationFrameLatencyDetail